Make sure panels do not need to link against libcommon.la
libcommon.la defines several new GTypes, so it cannot be loaded into the
same process multiple times.

Copy gconf-property-editor to libgnome-control-center as this is useful
for all panels. However, it will need to be converted to GSettings in the
near future.

Copy capplet-stock-icons to the mouse panel, as it is not used in any other
panels.

Remove references to functions not yet implemented for panels (help,
window icon, etc).