shell: make the .desktop file pass "--overview"
Opening the shell via the main gnome-control-center.desktop file
should display it in overview mode. In particular, if the shell is
already open, then "launching" it again via the main .desktop file
should switch the existing window to the overview rather than just
raising it but leaving it pointing to whatever panel it was pointing
to before.

https://bugzilla.gnome.org/show_bug.cgi?id=639180