background: Make icon list smaller on netbooks
The default treeview is quite tall by default, and with the addition
of the shell's toolbar, the titlebar and the GNOME shell panel,
the background panel would be too tall for netbook computers.

This works around the issue by setting a smaller height request
for the icon list when the screen on which the panel will be shown
are no taller than 768 pixels.

https://bugzilla.gnome.org/show_bug.cgi?id=645649