shell: Don't update the visibility of the lock button
Currently the lock button's visibility is updated on permission
changes - if the active panel does provide a permission object,
the button is shown, otherwise hidden.
As GtkLockButton itself does manage its visibility, it is not
necessary to handle it in g-c-c. In fact it is harmful, as the
more sophisticated handling in GtkLockButton is overwritten: for
instances, GtkLockButton hides itself if an action is allowed and
the permission can not be released. By overwriting the behavior,
gnome-control-center ends up showing a non-functional button.

https://bugzilla.gnome.org/show_bug.cgi?id=652299