lib: Disallow external panels from existing
Seeing as people were putting all kind of junk in it,
instead of being sensible, let's close the doors on
external panels.

If you work-around this in any way, you better have a
very good reason to, such as wanting to deny yourself
any support from the upstream gnome-control-center developers,
or wanting ridicule for your poor judgement in UI and code design.