color: Only allow the user to import custom profiles with colord >= 0.1.12
When I was adding the color panel, we ripped out the import code as it was all
pretty hacky and hardcoded, with the intention of doing it the right way in
colord rather than in each client.

At the moment if the user selects 'Import' then the panel does nothing as the
code that would actually make it work was missing.

I've now added the required client code in colord 0.1.12 which makes it possible
to actually import files not listed in the dropdown.

This makes it possible to install a display profile that's been downloaded into
the 'Downloads' directory by firefox.