wacom: Implement g-c-c calibration functionality
Adds and handles callback from the "Calibrate" button.
When clicked, we get the current calibration and feed
it in to the calibration utility. After it completes
we get back a modified calibration and an indication
of success. If it was successful we apply the changes,
if not, we ignore the modifed calibration.

https://bugzilla.gnome.org/show_bug.cgi?id=657423