wacom: split the main page controls in a separate GtkGrid
The effect we want to achieve is to have the model name and button box
padded 16px to the left of the tablet controls below.
Pack all the tablet controls in a separate GtkGrid, so we can just set a
margin on that and offset its contents from the model above.

https://bugzilla.gnome.org/show_bug.cgi?id=668881