mouse: Don't enable two-finger scrolling based on width
Reverts part of 109b36919921d757b93b91e8082620078608967e

Good idea in principle, albeit the hardware gets in our way. In
synaptics 1.4 two-finger emulation was enabled by default on devices
that reported finger width but had to since revert this. The hardware
doesn't really separate width and pressure, so a hard tap off a finger
gives a high width, a wide finger gives a high pressure on light
touches. We got spurious right-click presses when two-finger
emulation was enabled.

Now we have the control-center providing a checkbox for two-finger
emulation but it doesn't enable emulation, which to the users appears as if
two-finger scrolling is broken
(https://bugzilla.redhat.com/show_bug.cgi?id=738123).

We now disable the checkbox again for hardware that doesn't do two-fingers.
it's not reliable and we're likely to see further bugs when the spurious right
clicks take effect

https://bugzilla.gnome.org/show_bug.cgi?id=661963