shell: Fix resetting page selection
When the page selection happens when another page selection hasn't
started yet, we need to cancel the previous animation, or only one
of them might be effectively acted upon.

This fixes the page from being selected if the search term is changed
quickly (you'd end up at the overview page instead of staying on the
search results page).

Requires the clutter patch from:
https://bugzilla.gnome.org/show_bug.cgi?id=676334

https://bugzilla.gnome.org/show_bug.cgi?id=676328