network: Split the proxy parts from the .c and .ui file
A source file with 4k lines of code is not understandable, and the glade file
was becoming quite a challenge for glade.
To add more features we need to split things up just to keep them maintainable.

This is the first patch that just splits out the proxy bits. Other device types
will follow in the next few days.