user-accounts: port to GtkLevelBar
This also adds another out parameter to pw_strength(), which is a discrete
indication of the password strength (to be used with GtkLevelBar). The
advantage of having such a parameter back from pw_strength() is that we
can make sure the fill value and the hint string change at the same
time, for better consistency.

https://bugzilla.gnome.org/show_bug.cgi?id=679984