user-accounts: Fix match icon not updating in some cases
Check the new password matching after focusing out either entries.

Validate the content of the verify password entry after
focusing out either the new password entry or the verify entry.

Previously the user could mistype his new password in the new
password entry, but when it was corrected, the verify entry
would still be indicating that the confirmation password did not
match when it did.

https://bugzilla.gnome.org/show_bug.cgi?id=668844