system info: set related mime types when changing preferred apps
Setting the preferred music or video app only changed the handler
for ogg audio/video, leaving an inconsistent state where the previous
(system default) app would be used for other images, audio and video
formats.
Instead make it so that when the user changes his preferences,
the application becomes the default for all mime types in the same
group that it supports.
Nothing changes if configuration is modified externally (e.g.
by hand or through nautilus): expose as preferred app the one that
is configured to handle the primary content type.

https://bugzilla.gnome.org/show_bug.cgi?id=674084