bluetooth: Don't break the custom widgets on connect
We used to destroy and recreate the custom widgets whenever any of
the properties changed. Now we make sure that the custom widgets are
only destroyed and recreated when the device selected is a different one.

This fixes NetworkManager's Bluetooth plugins getting destroyed
because the Connected property changed, as it was trying to connect to
the device.

https://bugzilla.gnome.org/show_bug.cgi?id=681456