bluetooth: Fix main switch acting bizarrely
The only time we were looking at the powered state of the adapter
is when the killswitch state was changing. Except that we're fast enough
that bluetoothd didn't have time to power up the adapter, so its
state was unpowered, which we would set the switch to.

The switch was off, the adapter was on.

We fix that problem by tracking the powered state of the adapter
separately.

https://bugzilla.redhat.com/show_bug.cgi?id=841881